// abis/types/VOCVendor.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from "bn.js";
import {ContractOptions} from "web3-eth-contract";
import {EventLog} from "web3-core";
import {EventEmitter} from "events";
import {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export type TokensPurchased = ContractEventLog<{
  purchaser: string;
  beneficiary: string;
  value: string;
  amount: string;
  0: string;
  1: string;
  2: string;
  3: string;
}>;

export interface VOCVendor extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): VOCVendor;
  clone(): VOCVendor;
  methods: {
    /**
     * Allow users to buy token for ETH
     */
    buyTokens(beneficiary: string): PayableTransactionObject<void>;

    /**
     */
    getRate(): NonPayableTransactionObject<string>;

    /**
     */
    getToken(): NonPayableTransactionObject<string>;

    /**
     */
    getWallet(): NonPayableTransactionObject<string>;
  };
  events: {
    TokensPurchased(cb?: Callback<TokensPurchased>): EventEmitter;
    TokensPurchased(
      options?: EventOptions,
      cb?: Callback<TokensPurchased>
    ): EventEmitter;

    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };

  once(event: "TokensPurchased", cb: Callback<TokensPurchased>): void;
  once(
    event: "TokensPurchased",
    options: EventOptions,
    cb: Callback<TokensPurchased>
  ): void;
}